chore: split cli package into focused files
Reduce the size and scope of the monolithic cli entrypoint so command behavior is easier to navigate and maintain.

Add focused tests around the newly extracted image, policy, status, serve, and local image import paths so the refactor is covered by package-level behavior checks.
