+++ internal/cli/logging.go
@@ -1,0 +1,22 @@
+package cli
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/charmbracelet/log"
+)
+
+func newLogger(rawLevel, component string) (*log.Logger, error) {
+	levelName := effectiveLogLevel(rawLevel)
+	level, err := log.ParseLevel(levelName)
+	if err != nil {
+		return nil, fmt.Errorf("invalid --log-level %q: %w", rawLevel, err)
+	}
+	logger := log.NewWithOptions(os.Stderr, log.Options{
+		Level:     level,
+		Formatter: log.TextFormatter,
+	})
+	applyPolishedLoggerStyles(logger, shouldUseANSI(os.Stderr))
+	return logger.With("component", component), nil
+}
